✨ Touch Ups
Make use of all four characters that we have, we use our background song in our game state as well and we add the possibility to have sounds playing with your actions. We also make a small little entry for our player so you quickly see who you are playing as.